270_nsp/agent_bing_custom_search/requirements.txt
agent-framework==1.0.0b251223
agent-framework-azure-ai==1.0.0b251223
azure-ai-agentserver-agentframework==1.0.0b7
azure-ai-agentserver-core==1.0.0b7

python-dotenv>=1.0.0
azure-identity

# azure-monitor-opentelemetry==1.8.4

# azure-ai-agentserver-agentframework==1.0.0b7
# azure-ai-agentserver-core==1.0.0b7
# agent-framework==1.0.0b260116
# agent-framework-azure-ai==1.0.0b260116
# python-dotenv>=1.1.1
# azure-monitor-opentelemetry==1.8.1
